refactor: Migrate feature flags to annotation-based codegen
* Replace manual Metro DI bindings in `FlagBindings.kt` with the `@FeatureFlag` annotation to leverage code generation for qualifiers and bindings.
* Migrate `AccountSwitchFlag`, `ContinueWatchingNitroFlag`, and `SimklLoginFlag` to the new annotation-driven pattern.
* Add `EnableLiquidGlassFlag` using the `@FeatureFlag` platform parameter to scope the flag specifically to iOS.
* Move `DefaultFeatureFlagFactory` from the implementation package to the API package to support the updated architecture.
* Remove obsolete manual qualifier annotations for all migrated flags.
