+++ core/feature-flags/api/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 scaffold {
     useMetro()
+    useFeatureFlagCodegen()
 }
 
 kotlin {
