--- core/feature-flags/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/featureflags/flags/AccountSwitchFlagQualifier.kt
@@ -1,6 +1,0 @@
-package com.thomaskioko.tvmaniac.featureflags.flags
-
-import dev.zacsweers.metro.Qualifier
-
-@Qualifier
-public annotation class AccountSwitchFlagQualifier
